moved dependency utils into flow_integration_sdk
